+++ sleap_mot/utils.py
@@ -1,19 +1,155 @@
 """Helper functions for Tracker module."""
 
-from typing import List, Tuple, Union
 from scipy.optimize import linear_sum_assignment
 
 import numpy as np
 import sleap_io as sio
 
 
-def hungarian_matching(cost_matrix: np.ndarray) -> List[Tuple[int, int]]:
+def compute_instance_area(points: np.ndarray) -> np.ndarray:
+    """Compute the area of the bounding box of a set of keypoints.
+
+    Args:
+        points: A numpy array of coordinates.
+
+    Returns:
+        The area of the bounding box of the points.
+    """
+    if points.ndim == 2:
+        points = np.expand_dims(points, axis=0)
+
+    min_pt = np.nanmin(points, axis=-2)
+    max_pt = np.nanmax(points, axis=-2)
+
+    return np.prod(max_pt - min_pt, axis=-1)
+
+
+def compute_oks(
+    points_gt: np.ndarray,
+    points_pr: np.ndarray,
+    scale: float | None = None,
+    stddev: float = 0.025,
+    use_cocoeval: bool = True,
+) -> np.ndarray:
+    """Compute the object keypoints similarity between sets of points.
+
+    Args:
+        points_gt: Ground truth instances of shape (n_gt, n_nodes, n_ed),
+            where n_nodes is the number of body parts/keypoint types, and n_ed
+            is the number of Euclidean dimensions (typically 2 or 3). Keypoints
+            that are missing/not visible should be represented as NaNs.
+        points_pr: Predicted instance of shape (n_pr, n_nodes, n_ed).
+        use_cocoeval: Indicates whether the OKS score is calculated like cocoeval
+            method or not. True indicating the score is calculated using the
+            cocoeval method (widely used and the code can be found here at
+            https://github.com/cocodataset/cocoapi/blob/8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9/PythonAPI/pycocotools/cocoeval.py#L192C5-L233C20)
+            and False indicating the score is calculated using the method exactly
+            as given in the paper referenced in the Notes below.
+        scale: Size scaling factor to use when weighing the scores, typically
+            the area of the bounding box of the instance (in pixels). This
+            should be of the length n_gt. If a scalar is provided, the same
+            number is used for all ground truth instances. If set to None, the
+            bounding box area of the ground truth instances will be calculated.
+        stddev: The standard deviation associated with the spread in the
+            localization accuracy of each node/keypoint type. This should be of
+            the length n_nodes. "Easier" keypoint types will have lower values
+            to reflect the smaller spread expected in localizing it.
+
+    Returns:
+        The object keypoints similarity between every pair of ground truth and
+        predicted instance, a numpy array of of shape (n_gt, n_pr) in the range
+        of [0, 1.0], with 1.0 denoting a perfect match.
+
+    Notes:
+        It's important to set the stddev appropriately when accounting for the
+        difficulty of each keypoint type. For reference, the median value for
+        all keypoint types in COCO is 0.072. The "easiest" keypoint is the left
+        eye, with stddev of 0.025, since it is easy to precisely locate the
+        eyes when labeling. The "hardest" keypoint is the left hip, with stddev
+        of 0.107, since it's hard to locate the left hip bone without external
+        anatomical features and since it is often occluded by clothing.
+
+        The implementation here is based off of the descriptions in:
+        Ronch & Perona. "Benchmarking and Error Diagnosis in Multi-Instance Pose
+        Estimation." ICCV (2017).
+    """
+    if points_gt.ndim == 2:
+        points_gt = np.expand_dims(points_gt, axis=0)
+    if points_pr.ndim == 2:
+        points_pr = np.expand_dims(points_pr, axis=0)
+
+    if scale is None:
+        scale = compute_instance_area(points_gt)
+
+    n_gt, n_nodes, n_ed = points_gt.shape  # n_ed = 2 or 3 (euclidean dimensions)
+    n_pr = points_pr.shape[0]
+
+    # If scalar scale was provided, use the same for each ground truth instance.
+    if np.isscalar(scale):
+        scale = np.full(n_gt, scale)
+
+    # If scalar standard deviation was provided, use the same for each node.
+    if np.isscalar(stddev):
+        stddev = np.full(n_nodes, stddev)
+
+    # Compute displacement between each pair.
+    displacement = np.reshape(points_gt, (n_gt, 1, n_nodes, n_ed)) - np.reshape(
+        points_pr, (1, n_pr, n_nodes, n_ed)
+    )
+    assert displacement.shape == (n_gt, n_pr, n_nodes, n_ed)
+
+    # Convert to pairwise Euclidean distances.
+    distance = (displacement**2).sum(axis=-1)  # (n_gt, n_pr, n_nodes)
+    assert distance.shape == (n_gt, n_pr, n_nodes)
+
+    # Compute the normalization factor per keypoint.
+    if use_cocoeval:
+        # If use_cocoeval is True, then compute normalization factor according to cocoeval.
+        spread_factor = (2 * stddev) ** 2
+        scale_factor = 2 * (scale + np.spacing(1))
+    else:
+        # If use_cocoeval is False, then compute normalization factor according to the paper.
+        spread_factor = stddev**2
+        scale_factor = 2 * ((scale + np.spacing(1)) ** 2)
+    normalization_factor = np.reshape(spread_factor, (1, 1, n_nodes)) * np.reshape(
+        scale_factor, (n_gt, 1, 1)
+    )
+    assert normalization_factor.shape == (n_gt, 1, n_nodes)
+
+    # Since a "miss" is considered as KS < 0.5, we'll set the
+    # distances for predicted points that are missing to inf.
+    missing_pr = np.any(np.isnan(points_pr), axis=-1)  # (n_pr, n_nodes)
+    assert missing_pr.shape == (n_pr, n_nodes)
+    distance[:, missing_pr] = np.inf
+
+    # Compute the keypoint similarity as per the top of Eq. 1.
+    ks = np.exp(-(distance / normalization_factor))  # (n_gt, n_pr, n_nodes)
+    assert ks.shape == (n_gt, n_pr, n_nodes)
+
+    # Set the KS for missing ground truth points to 0.
+    # This is equivalent to the visibility delta function of the bottom
+    # of Eq. 1.
+    missing_gt = np.any(np.isnan(points_gt), axis=-1)  # (n_gt, n_nodes)
+    assert missing_gt.shape == (n_gt, n_nodes)
+    ks[np.expand_dims(missing_gt, axis=1)] = 0
+
+    # Compute the OKS.
+    n_visible_gt = np.sum(
+        (~missing_gt).astype("float32"), axis=-1, keepdims=True
+    )  # (n_gt, 1)
+    oks = np.sum(ks, axis=-1) / n_visible_gt
+    assert oks.shape == (n_gt, n_pr)
+
+    return oks
+
+
+def hungarian_matching(cost_matrix: np.ndarray) -> list[tuple[int, int]]:
     """Match new instances to existing tracks using Hungarian matching."""
     row_ids, col_ids = linear_sum_assignment(cost_matrix)
     return row_ids, col_ids
 
 
-def greedy_matching(cost_matrix: np.ndarray) -> List[Tuple[int, int]]:
+def greedy_matching(cost_matrix: np.ndarray) -> list[tuple[int, int]]:
     """Match new instances to existing tracks using greedy bipartite matching."""
     # Sort edges by ascending cost.
     rows, cols = np.unravel_index(np.argsort(cost_matrix, axis=None), cost_matrix.shape)
@@ -35,14 +171,14 @@ def greedy_matching(cost_matrix: np.ndarray) -> List[Tuple[int, int]]:
     return row_inds, col_inds
 
 
-def get_keypoints(pred_instance: Union[sio.PredictedInstance, np.ndarray]):
+def get_keypoints(pred_instance: sio.PredictedInstance | np.ndarray):
     """Return keypoints as np.array from the `PredictedInstance` object."""
     if isinstance(pred_instance, np.ndarray):
         return pred_instance
     return pred_instance.numpy()
 
 
-def get_centroid(pred_instance: Union[sio.PredictedInstance, np.ndarray]):
+def get_centroid(pred_instance: sio.PredictedInstance | np.ndarray):
     """Return the centroid of the `PredictedInstance` object."""
     pts = pred_instance
     if not isinstance(pred_instance, np.ndarray):
@@ -51,7 +187,7 @@ def get_centroid(pred_instance: Union[sio.PredictedInstance, np.ndarray]):
     return centroid
 
 
-def get_bbox(pred_instance: Union[sio.PredictedInstance, np.ndarray]):
+def get_bbox(pred_instance: sio.PredictedInstance | np.ndarray):
     """Return the bounding box coordinates for the `PredictedInstance` object."""
     points = (
         pred_instance.numpy()
